refactor: move database related components
